ecommerce/pyproject.toml
[tool.poetry]
name = "drf ecommers"
version = "0.1.0"
description = ""
authors = ["shivam <jhodshivam5122@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
pytest = "^7.3.1"
pytest-django = "^4.5.2"
black = "^23.3.0"
django-mptt = "^0.14.0"
drf-spectacular = "^0.26.1"
coverage = "^7.2.3"
pytest-cov = "^4.0.0"
pytest-factoryboy = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
